Update bootstrap.js
ok now on open its showing on second monitor based on my multi mon ubuntu test. it was very fishy, if screenX and screenY were 0 0 then the first moveTo call would not do anyhting so it would stay on the screen it opened from. so i set screenX and Y to 1 1 then the moveTo actually moves it. also i did some testing with gtk_window_present and ia it seems that it ensures that when it shows editor window on one, it will also present it on others (recall i had issue where it would show on one monitor, and it was expanded, but it wouldnt show it on the other monitor until i clicked back into the one showing in monitor 1)